docusaurus: fix citation regression
