+++ src/resources/extensions/quarto/docusaurus/_extension.yml
@@ -20,8 +20,8 @@ contributes:
       # Although we use a custom writer, we still need the variants here the lua filters to render correctly.
       # Ideally, we would forward the variants to the custom writer.
       variant: gfm+pipe_tables+tex_math_dollars+header_attributes+raw_html+all_symbols_escapable+backtick_code_blocks+fenced_code_blocks+space_in_atx_header+intraword_underscores+lists_without_preceding_blankline+shortcut_reference_links
-      writer: docusaurus_writer.lua
       output-ext: mdx
+      quarto-custom-format: docusaurus
       inline-includes: true
       preserve-yaml: true
       wrap: none
@@ -30,6 +30,7 @@ contributes:
       fig-height: 5
       html-math-method: webtex
       filters:
-        - docusaurus.lua
-  shortcodes: # this is a hack, we just want to include this file
-    - docusaurus_renderers.lua
+        - at: post-quarto
+          path: docusaurus.lua
+        - at: post-finalize
+          path: docusaurus_citeproc.lua
