+++ tests/docs/smoke-all/2023/04/27/5316.qmd
@@ -7,8 +7,8 @@ _quarto:
   tests:
     docusaurus-md:
       ensureFileRegexMatches:
-        - ["[^\\]]```python", "[^\\]]```mdx-code-block"]
-        - ["[^\\]]```text"]
+        - ["[^\\]]``` *python", "[^\\]]``` *mdx-code-block"]
+        - ["[^\\]]``` *text"]
 ---
 
 ```{.python title="Python"}
